requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
aniso8601==7.0.0
appnope==0.1.2
asgiref==3.4.1
astroid==2.7.3
async-timeout==4.0.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
decorator==5.1.0
Django==3.2.7
django-address==0.2.5
django-cors-headers==3.8.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-grappelli==2.15.1
django-model-utils==4.1.1
django-money==2.0.3
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
frozenlist==1.2.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
hashids==1.2.0
idna==3.3
importlib-metadata==4.8.2
ipython==7.29.0
isort==5.9.3
jedi==0.18.0
k2-connect==1.2.2
lazy-object-proxy==1.6.0
Markdown==3.3.6
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.31
pickleshare==0.7.5
platformdirs==2.3.0
promise==2.3
prompt-toolkit==3.0.22
psycopg2==2.9.1
ptyprocess==0.7.0
py-moneyed==1.2
Pygments==2.10.0
PyJWT==1.7.1
pylint==2.10.2
python-dotenv==0.19.1
pytz==2021.1
regex==2021.8.28
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
slack-sdk==3.11.2
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0
